[TIC-34] fix: Minor issues fixing
